Debug unit tests

+++ core/src/test/java/com/github/drinkjava2/functionstest/BatchTest.java
@@ -11,6 +11,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.github.drinkjava2.config.DataSourceConfig.H2DataSourceBox;
 import com.github.drinkjava2.config.DataSourceConfig.MySqlDataSourceBox;
 import com.github.drinkjava2.jbeanbox.BeanBox;
 import com.github.drinkjava2.jdialects.annotation.jpa.Table;
@@ -24,11 +25,14 @@ import com.github.drinkjava2.jsqlbox.SqlBoxContext;
  * @since 1.7.0
  */
 public class BatchTest {
-	static {
-		SqlBoxContext.setGlobalAllowShowSql(true);
-	}
-	SqlBoxContext ctx = new SqlBoxContext((DataSource) BeanBox.getBean(MySqlDataSourceBox.class));
+	int repeat = 5000;
+
+	SqlBoxContext ctx = null;
+
 	{
+		// SqlBoxContext.setGlobalAllowShowSql(true);
+		// Need change to MySqlDataSourceBox.class because H2 does not support batch
+		ctx = new SqlBoxContext((DataSource) BeanBox.getBean(H2DataSourceBox.class));
 		SqlBoxContext.setGlobalSqlBoxContext(ctx);
 	}
 
@@ -68,43 +72,49 @@ public class BatchTest {
 
 	@Test
 	public void testBatch() {
-		int repeat = 1000;
-
+		if (!ctx.getDialect().isMySqlFamily()) {
+			System.out.println("This test only can run on databases support batch insert/update");
+			return;
+		}
 		User user = new User();
 		user.setName("Sam");
 		user.setAddress("Canada");
 
+		System.out.println("=======================Normal insert begin=====================");
 		ctx.nExecute("delete from batch_test_tb");
-		
-		
 		long start = System.currentTimeMillis();
 		for (long i = 0; i < repeat; i++) {
 			user.setName("Name" + i);
 			user.setAddress("Address" + i);
 			user.insert();
 		}
-
 		long end = System.currentTimeMillis();
 		String timeused = "" + (end - start) / 1000 + "." + (end - start) % 1000;
 		System.out.println(String.format("Non-Batch execute " + repeat + " SQLs time used: %6s s", timeused));
 		Assert.assertEquals(repeat, ctx.nQueryForLongValue("select count(*) from batch_test_tb"));
+		System.out.println("=======================Normal insert end=====================\n\n");
 
+		System.out.println("=======================nBatchBegin() =====================");
 		ctx.nExecute("delete from batch_test_tb");
 		start = System.currentTimeMillis();
-		ctx.nBatchBegin();
-		for (long i = 0; i < repeat; i++) {
-			user.setName("Name" + i);
-			user.setAddress("Address" + i);
-			user.insert();
+		try {
+			ctx.nBatchBegin();
+			for (long i = 0; i < repeat; i++) {
+				user.setName("Name" + i);
+				user.setAddress("Address" + i);
+				user.insert();
+			}
+		} finally {
+			ctx.nBatchEnd();
 		}
-
-		ctx.nBatchEnd();
 		end = System.currentTimeMillis();
 		timeused = "" + (end - start) / 1000 + "." + (end - start) % 1000;
 		System.out
 				.println(String.format("nBatchBegin/nBatchEnd execute " + repeat + " SQLs time used: %6s s", timeused));
 		Assert.assertEquals(repeat, ctx.nQueryForLongValue("select count(*) from batch_test_tb"));
+		System.out.println("=======================nBatchEnd() =====================\n\n");
 
+		System.out.println("=======================nBatch method begin=====================");
 		ctx.nExecute("delete from batch_test_tb");
 		start = System.currentTimeMillis();
 		List<Object[]> params = new ArrayList<Object[]>();
@@ -116,7 +126,9 @@ public class BatchTest {
 		System.out.println(String
 				.format("nBatch(Sql, List<Object[])) method execute " + repeat + " SQLs time used: %6s s", timeused));
 		Assert.assertEquals(repeat, ctx.nQueryForLongValue("select count(*) from batch_test_tb"));
+		System.out.println("=======================nBatch method end=====================\n\n");
 
+		System.out.println("=======================DbUtils batch method begin=====================");
 		ctx.nExecute("delete from batch_test_tb");
 		start = System.currentTimeMillis();
 		Object[][] paramsArray = new Object[repeat][2];
@@ -134,6 +146,7 @@ public class BatchTest {
 		System.out.println(
 				String.format("batch(Sql, Object[][]) method execute " + repeat + " SQLs time used: %6s s", timeused));
 		Assert.assertEquals(repeat, ctx.nQueryForLongValue("select count(*) from batch_test_tb"));
+		System.out.println("=======================DbUtils batch method end=====================\n\n");
 	}
 
 }
